joystick: rename file names
