+++ template_joystick_control/template_joystick_control/task_basin.py
@@ -7,7 +7,7 @@ from typing import Tuple, Optional
 from std_msgs.msg import Float32MultiArray
 
 from template_joystick_control.channel import Channel
-from template_joystick_control.joystick_mapping import JoystickAxes
+from template_joystick_control.mapping import JoystickAxes
 from numpy.typing import ArrayLike
