+++ template_joystick_control/template_joystick_control/task_simple.py
@@ -6,7 +6,7 @@ import numpy as np
 from typing import Tuple
 from numpy.typing import ArrayLike
 
-from template_joystick_control.joystick_mapping import JoystickAxes
+from template_joystick_control.mapping import JoystickAxes
 from template_joystick_control.channel import Channel
